+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Ported `settings` to `Arduino` framework;
 * Added `DEBUG` flag to interact with `sim7070g` module;
 * Ported `devices` to `Arduino` framework;
+* Ported `rs485` module to `Arduino` framework;
 
 ### Changed